requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.8
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jsonpath-python==1.0.6
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
milvus-lite==2.4.11
mistralai==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
ollama==0.4.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.4
PyMuPDF==1.25.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.34.0
wcwidth==0.2.13
